am62ax: UDMA: Fix Misra C Rule 10.7 issue
MISRA.ETYPE.COMP.CAST.IMPL.WIDER

A composite expression used as an operand of an operator in which the
usual arithmetic conversions are performed has its other operand
having a wider essential type.

Fixes: SITSW-7048

+++ source/drivers/udma/v0/udma_ch.c
@@ -4189,7 +4189,7 @@ static void Udma_chSetPeerAasrcReg(Udma_DrvHandleInt drvHandle,
              CSL_FMK(PSILCFG_REG_FIFOCFG_DMA_REQ_RESET, (uint32_t)pdmaPrms->dmaReqReset) |
              CSL_FMK(PSILCFG_REG_FIFOCFG_LAST_SLOT, (uint32_t)pdmaPrms->lastSlot) |
              CSL_FMK(PSILCFG_REG_FIFOCFG_FIRST_SLOT, (uint32_t)pdmaPrms->firstSlot) |
-             CSL_FMK(PSILCFG_REG_FIFOCFG_DMA_REQ_MASK, pdmaPrms->dmaReqMask);
+             CSL_FMK(PSILCFG_REG_FIFOCFG_DMA_REQ_MASK, (uint32_t)pdmaPrms->dmaReqMask);
     CSL_REG32_WR(PEER5, regVal);
 
     DebugP_assert(PEER6 != NULL);
